tweak swift-format; add app
