+++ LuasWatch Watch App/Coordinator.swift
@@ -3,9 +3,9 @@
 //  Copyright © 2019 mApps.ie. All rights reserved.
 //
 
+import Combine
 import Foundation
 import LuasKit
-import Combine
 
 class Coordinator: NSObject {
 
@@ -18,8 +18,10 @@ class Coordinator: NSObject {
     private static let refreshInterval = 12.0
     private var cancellable: AnyCancellable?
 
-	init(appModel: AppModel,
-		 location: Location) {
+    init(
+        appModel: AppModel,
+        location: Location
+    ) {
         self.appModel = appModel
         self.location = location
         self.cancellable = appModel.$appState
@@ -55,10 +57,11 @@ class Coordinator: NSObject {
         #warning("a bit ugly -  notification is sent by ChangeStationButton - is there a better way?")
         NotificationCenter.default.addObserver(
             forName: Notification.Name("LuasWatch.RetriggerTimer"),
-            object: nil, queue: nil) { _ in
-                self.retriggerTimer()
-            }
-	}
+            object: nil, queue: nil
+        ) { _ in
+            self.retriggerTimer()
+        }
+    }
 
     func invalidateTimer() {
         timer?.invalidate()
@@ -77,39 +80,46 @@ class Coordinator: NSObject {
         timerDidFire()
 
         // ... and then schedule again for regular interval
-        timer = Timer.scheduledTimer(timeInterval: Self.refreshInterval,
-                                     target: self, selector: #selector(timerDidFire),
-                                     userInfo: nil, repeats: true)
+        timer = Timer.scheduledTimer(
+            timeInterval: Self.refreshInterval,
+            target: self, selector: #selector(timerDidFire),
+            userInfo: nil, repeats: true)
     }
 
-	@objc func timerDidFire() {
+    @objc func timerDidFire() {
         #warning("check whether modal is up")
-        
-		// if user has selected a specific station
+
+        // if user has selected a specific station
         if let station = appModel.appMode.specificStation {
 
             // the location we have is not too old -> don't wait for another location update
             if let latestLocation = appModel.latestLocation,
-               latestLocation.isQuiteRecent() {
+                latestLocation.isQuiteRecent()
+            {
                 myPrint("🥳 we have user selected station & recent location -> skip location update")
                 handle(station, latestLocation)
             } else {
-                myPrint("😇 user has selected specific station & only outdated or no location \(appModel.latestLocation?.timestamp.timeIntervalSinceNow ?? 0) -> wait for location update")
+                myPrint(
+                    "😇 user has selected specific station & only outdated or no location \(appModel.latestLocation?.timestamp.timeIntervalSinceNow ?? 0) -> wait for location update"
+                )
                 location.update()
             }
 
         } else {
-			// user has NOT selected a specific station
+            // user has NOT selected a specific station
 
             if let latestLocation = appModel.latestLocation,
-               latestLocation.isQuiteRecent() {
+                latestLocation.isQuiteRecent()
+            {
                 // we have a location that's not too old
                 myPrint("🥳 user has NOT selected specific station & we have a recent location -> skip location update")
                 didGetLocation(latestLocation)
             } else {
-                myPrint("😇 user has NOT selected specific station & only outdated location \(appModel.latestLocation?.timestamp.timeIntervalSinceNow ?? 0) -> wait for location update")
+                myPrint(
+                    "😇 user has NOT selected specific station & only outdated location \(appModel.latestLocation?.timestamp.timeIntervalSinceNow ?? 0) -> wait for location update"
+                )
                 location.update()
             }
-		}
-	}
+        }
+    }
 }
