+++ LuasWatch Watch App/Views/LinesView.swift
@@ -3,8 +3,8 @@
 //  Copyright © 2024 mApps.ie. All rights reserved.
 //
 
-import SwiftUI
 import LuasKit
+import SwiftUI
 
 struct LinesView: View {
     var actionGreen: () -> Void
