+++ LuasWatch Watch App/Views/OLD/TapOverlayView.swift
@@ -5,7 +5,7 @@
 
 import SwiftUI
 
-extension LuasView {
+extension LuasViewOLD {
 
     @ViewBuilder
     internal func tapOverlayView() -> some View {
