+++ LuasWatch.xcodeproj/xcshareddata/xcschemes/LuasWatchComplicationsExtension.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1500"
+   LastUpgradeVersion = "1520"
    version = "2.0">
    <BuildAction
       parallelizeBuildables = "YES"
